Added auto calculated fields and SPARK UDFs (#1669)
* udf functions added in writer

* rename autocalculated fields and trigger validations

* change in models with field

* [SPARTA-695]: Add a first version of auto calculated field component

* [PAAS-1653]: Add auto calculated field component to triggers

* [SPARTA-695]: Adapt trigger model to new format

* [SPARTA-695]: Fix minor bug to form array component

* [SPARTA-695]: Add title to each auto calculated field when user is creating them

* [SPARTA-695]: Fix add button to array form

* added auxSchema to get metadata info in dataFrame modifier

* log fragments in debug mode

* [SPARTA-695]: Clean policy id when policy is initialized.

* correctly end in local mode

+++ driver/src/main/scala/com/stratio/sparta/driver/factory/SparkContextFactory.scala
@@ -125,15 +125,15 @@ object SparkContextFactory extends SLF4JLogging {
     synchronized {
       ssc.fold(log.warn("Spark Streaming Context is empty")) { streamingContext =>
         try {
-          val stopGracefully = Try(SpartaConfig.getDetailConfig.get.getBoolean(AppConstant.ConfigStopGracefully))
-            .getOrElse(true)
-          log.info(s"Stopping streamingContext with name: ${streamingContext.sparkContext.appName}")
-          Try(streamingContext.stop(false, stopGracefully)) match {
+          val stopGracefully =
+            Try(SpartaConfig.getDetailConfig.get.getBoolean(AppConstant.ConfigStopGracefully)).getOrElse(true)
+          log.info(s"Stopping Streaming Context with name: ${streamingContext.sparkContext.appName}")
+          Try(streamingContext.stop(stopSparkContext = false, stopGracefully)) match {
             case Success(_) =>
-              log.info("Streaming context have been stopped gracefully")
+              log.info("Streaming Context have been stopped gracefully")
             case Failure(error) =>
-              log.error("Streaming context is not stopped gracefully, now it will be stopped forcedly", error)
-              streamingContext.stop(false, false)
+              log.error("Streaming Context is not stopped gracefully, now it will be stopped forcibly", error)
+              streamingContext.stop(stopSparkContext = false, stopGracefully = false)
           }
         } finally {
           ssc = None
@@ -142,17 +142,20 @@ object SparkContextFactory extends SLF4JLogging {
     }
   }
 
-  def destroySparkContext(): Unit = {
-    synchronized {
+  def destroySparkContext(destroyStreamingContext: Boolean = true): Unit = {
+    if (destroyStreamingContext)
       destroySparkStreamingContext()
-      sc.fold(log.warn("Spark Context is empty")) { sparkContext =>
-        try {
-          log.info("Stopping SparkContext with name: " + sparkContext.appName)
-          sparkContext.stop()
-          log.info("Stopped SparkContext with name: " + sparkContext.appName)
-        } finally {
-          sc = None
-        }
+
+    sc.fold(log.warn("Spark Context is empty")) { sparkContext =>
+      try {
+        log.info("Stopping SparkContext with name: " + sparkContext.appName)
+        sparkContext.stop()
+        log.info("Stopped SparkContext with name: " + sparkContext.appName)
+      } finally {
+        sqlContext = None
+        sqlInitialSentences =  Seq.empty[String]
+        ssc = None
+        sc = None
       }
     }
   }
